Fix decoding of BASIC auth credentials
The decoding process is messing up passwords that contain a ":"
character. For example,
if a username/password combination like { "user", "abcd:1234" } 
is passed in, the password is interpreted as "abcd".

This commit fixes that problem. The first colon encountered in the auth
string is the only one that's interpreted. That is, in this commit, 
the password is correctly interpreted as "abcd:1234".